pyproject.toml
[project]
name = "egile-mcp-reporter"
version = "0.1.0"
description = "MCP server for generating professional reports in multiple formats (PDF, PowerPoint, HTML, Markdown)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Egile Team" }
]
keywords = ["mcp", "reports", "pdf", "powerpoint", "html", "markdown", "data-visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp",
    "fastmcp>=0.1.0",
    "python-pptx>=0.6.21",
    "fpdf2>=2.7.0",
    "Pillow>=10.0.0",
    "markdown>=3.5.0",
    "plotly>=5.18.0",
    "pandas>=2.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
egile-mcp-reporter = "egile_mcp_reporter.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/egile_mcp_reporter"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
